+++ cms/tests/lib/tearDownStrapi.ts
@@ -1,10 +1,13 @@
 import fs from 'fs';
 
+/**
+ * should only be called if setUpStrapi has been called
+ */
 export function teardownDatabase() {
 	const dbSettings = (strapi as any).config.get('database.connections.default.settings');
 
 	if (dbSettings && dbSettings.filename) {
-		// __dirname is /dsckiit-website-2.0/cms/tests/helpers
+		// __dirname is /dsckiit-website-2.0/cms/tests/lib
 		// dbSettings.filename is .tmp/test.db
 		const tmpDbFile = `${__dirname}/../../${dbSettings.filename}`;
 		if (fs.existsSync(tmpDbFile)) {
