+++ cms/tsconfig.json
@@ -39,7 +39,7 @@
 		// "noUncheckedIndexedAccess": true,            /* Include 'undefined' in index signature results */
 		// "noPropertyAccessFromIndexSignature": true,  /* Require undeclared properties from index signatures to use element accesses. */
 		/* Module Resolution Options */
-		// "moduleResolution": "node",                  /* Specify module resolution strategy: 'node' (Node.js) or 'classic' (TypeScript pre-1.6). */
+		"moduleResolution": "node" /* Specify module resolution strategy: 'node' (Node.js) or 'classic' (TypeScript pre-1.6). */,
 		// "baseUrl": "./",                             /* Base directory to resolve non-absolute module names. */
 		// "paths": {},                                 /* A series of entries which re-map imports to lookup locations relative to the 'baseUrl'. */
 		// "rootDirs": [],                              /* List of root folders whose combined content represents the structure of the project at runtime. */
